requirements.txt
fastapi 
uvicorn 
sqlalchemy 
alembic 
psycopg2-binary
slowapi 
python-dotenv 
pyjwt 
passlib 
bcrypt 
python-multipart 
email-validator